+++ src/Exceptions/AuthenticationException.php
@@ -7,23 +7,16 @@ use Illuminate\Auth\AuthenticationException as IlluminateAuthenticationException
 class AuthenticationException extends IlluminateAuthenticationException implements RendersErrorsExtensions
 {
     public const MESSAGE = 'Unauthenticated.';
-    public const CATEGORY = 'authentication';
 
     public function isClientSafe(): bool
     {
         return true;
     }
 
-    public function getCategory(): string
-    {
-        return self::CATEGORY;
-    }
-
-    /**
-     * @return array<string, array<string>>
-     */
     public function extensionsContent(): array
     {
-        return ['guards' => $this->guards];
+        return [
+            'guards' => $this->guards,
+        ];
     }
 }