+++ src/Exceptions/DefinitionException.php
@@ -16,9 +16,4 @@ class DefinitionException extends Exception implements ClientAware
     {
         return false;
     }
-
-    public function getCategory(): string
-    {
-        return 'schema';
-    }
 }