+++ src/Exceptions/DirectiveException.php
@@ -11,9 +11,4 @@ class DirectiveException extends Exception implements ClientAware
     {
         return false;
     }
-
-    public function getCategory(): string
-    {
-        return 'schema';
-    }
 }